chore: Separate type generator class
