+++ src/analyzer/analyzer.ts
@@ -2,15 +2,15 @@ import ts from 'typescript';
 import path from 'path';
 import { ScriptSourceHelper } from '../ts-ast-util/types';
 import { Extractor } from './extractor';
-import { createScriptSourceHelper, createSourceWriteHelper } from '../ts-ast-util';
+import { createScriptSourceHelper } from '../ts-ast-util';
 import { TsGraphQLPluginConfigOptions } from '../types';
 import { SchemaManager, SchemaBuildErrorInfo } from '../schema-manager/schema-manager';
 import { TsGqlError, ErrorWithLocation, ErrorWithoutLocation } from '../errors';
-import { location2pos, dasherize } from '../string-util';
+import { location2pos } from '../string-util';
 import { validate } from './validator';
 import { ManifestOutput } from './types';
 import { MarkdownReporter } from './markdown-reporter';
-import { TypeGenVisitor, TypeGenError } from '../typegen/type-gen-visitor';
+import { TypeGenerator } from './type-generator';
 
 export function convertSchemaBuildErrorsToErrorWithLocation(errorInfo: SchemaBuildErrorInfo) {
   const fileName = errorInfo.fileName;
@@ -31,6 +31,7 @@ export function convertSchemaBuildErrorsToErrorWithLocation(errorInfo: SchemaBui
 export class Analyzer {
   private _extractor: Extractor;
   private _scriptSourceHelper: ScriptSourceHelper;
+  private _typeGenerator: TypeGenerator;
 
   constructor(
     private readonly _pluginConfig: TsGraphQLPluginConfigOptions,
@@ -47,6 +48,11 @@ export class Analyzer {
     this._extractor = new Extractor({
       removeDuplicatedFragments: this._pluginConfig.removeDuplicatedFragments === false ? false : true,
       scriptSourceHelper: this._scriptSourceHelper,
+      debug: this._debug,
+    });
+    this._typeGenerator = new TypeGenerator({
+      prjRootPath: this._prjRootPath,
+      extractor: this._extractor,
       debug: this._debug,
     });
   }
@@ -106,60 +112,8 @@ export class Analyzer {
     if (extractedErrors.length) {
       this._debug(`Found ${extractedErrors.length} extraction errors.`);
     }
-    const typegenErrors: TsGqlError[] = [];
-    const visitor = new TypeGenVisitor({ schema });
-    const outputSourceFiles: { fileName: string; content: string }[] = [];
-    const printer = ts.createPrinter({ newLine: ts.NewLineKind.LineFeed, removeComments: false });
-    extractedResults.forEach(extractedResult => {
-      if (extractedResult.documentNode) {
-        const { type, fragmentName, operationName } = this._extractor.getDominantDefiniton(extractedResult);
-        if (type === 'complex') {
-          const fileName = extractedResult.fileName;
-          const content = extractedResult.templateNode.getSourceFile().getFullText();
-          const start = extractedResult.templateNode.getStart();
-          const end = extractedResult.templateNode.getEnd();
-          const errorContent = { fileName, content, start, end };
-          const error = new ErrorWithLocation('This document node has complex operations.', errorContent);
-          typegenErrors.push(error);
-          return;
-        }
-        const operationOrFragmentName = type === 'fragment' ? fragmentName : operationName;
-        if (!operationOrFragmentName) return;
-        const outputFileName = path.resolve(
-          path.dirname(extractedResult.fileName),
-          '__generated__',
-          dasherize(operationOrFragmentName) + '.ts',
-        );
-        try {
-          // TODO Create type gen addon and pass it visitor.
-          const sourceWriteHelper = createSourceWriteHelper({ outputFileName });
-          const outputSourceFile = visitor.visit(extractedResult.documentNode, { sourceWriteHelper });
-          const content = printer.printFile(outputSourceFile);
-          outputSourceFiles.push({ fileName: outputFileName, content });
-          this._debug(
-            `Create type source file '${path.relative(this._prjRootPath, outputFileName)}' from '${path.relative(
-              this._prjRootPath,
-              extractedResult.fileName,
-            )}'.`,
-          );
-        } catch (error) {
-          if (error instanceof TypeGenError) {
-            const sourcePosition = extractedResult.resolevedTemplateInfo.getSourcePosition(error.node.loc!.start);
-            if (sourcePosition.isInOtherExpression) return;
-            const fileName = extractedResult.fileName;
-            const content = extractedResult.templateNode.getSourceFile().getFullText();
-            const start = sourcePosition.pos;
-            const end = extractedResult.resolevedTemplateInfo.getSourcePosition(error.node.loc!.end).pos;
-            const errorContent = { fileName, content, start, end };
-            const translatedError = new ErrorWithLocation(error.message, errorContent);
-            typegenErrors.push(translatedError);
-          } else {
-            throw error;
-          }
-        }
-      }
-    });
-    return { errors: [...schemaErrors, ...extractedErrors, ...typegenErrors], outputSourceFiles };
+    const { errors, outputSourceFiles } = this._typeGenerator.generateTypes({ schema, extractedResults });
+    return { errors: [...schemaErrors, ...extractedErrors, ...errors], outputSourceFiles };
   }
 
   private async _getSchema() {
